+++ packages/core/src/customer/strategies/googlepay/googlepay-customer-strategy.ts
@@ -1,5 +1,7 @@
 import { FormPoster } from '@bigcommerce/form-poster';
 
+import { getShippableItemsCount } from '@bigcommerce/checkout-sdk/payment-integration-api';
+
 import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
 import {
     InvalidArgumentError,
@@ -11,7 +13,6 @@ import { SDK_VERSION_HEADERS } from '../../../common/http-request';
 import { bindDecorator as bind } from '../../../common/utility';
 import { GooglePayPaymentProcessor } from '../../../payment/strategies/googlepay';
 import { RemoteCheckoutActionCreator } from '../../../remote-checkout';
-import { getShippableItemsCount } from '../../../shipping';
 import {
     CustomerInitializeOptions,
     CustomerRequestOptions,